requirements.txt
torch>=2.2.0
transformers>=4.41.0
datasets>=2.19.0
accelerate>=0.30.0
fastapi>=0.110.0
uvicorn>=0.29.0
Pillow>=10.0.0
scikit-learn>=1.3.0
